+++ src/lua/lua_api.cpp
@@ -1925,6 +1925,12 @@ void load_api(LuaEnvironment& env)
 			stack.push_float(t);
 			return 1;
 		});
+	env.add_module_function("RenderWorld", "mesh_set_material", [](lua_State* L)
+		{
+			LuaStack stack(L);
+			stack.get_render_world(1)->mesh_set_material(stack.get_mesh_instance(2), stack.get_string_id_64(3));
+			return 0;
+		});
 	env.add_module_function("RenderWorld", "mesh_set_visible", [](lua_State* L)
 		{
 			LuaStack stack(L);