+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Changelog
 ------
 *DD MMM YYYY*
 
+* lua: added RenderWorld.mesh_set_material()
 * tools: fixed a crash when entering empty commands in the console
 * tools: fixed an issue that prevented some operations in the Level Editor from being (un/re)done
 * tools: fixed an issue that prevented the data compiler from restoring and saving its state when launched by the Level Editor